Moved the validation in the config section
