+++ config/validators/index.js
@@ -1,0 +1,11 @@
+const { validationResult } = require('express-validator');
+
+const validationMid = (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
+  return next();
+};
+
+module.exports = validationMid;
